+++ SimpleEngine/SimpleEngine/Vector2.h
@@ -12,6 +12,6 @@ struct Vector2
 	Vector2 operator-();
 	Vector2 operator+(const Vector2 vec1);
 	Vector2 operator-(const Vector2 vec1);
-	Vector2 operator*(const double a);
+	Vector2 operator*(const float a);
 };
